AttackedWebsite/src/backend/requirements.txt
auto_mix_prep==0.2.0
Flask==3.0.3
Flask-Cors==5.0.0
psycopg2==2.9.9
PyYAML==6.0.2
xmltodict==0.13.0
